Ability to reconnect gracefully during server restarts. (beta)
http://discuss.dev.twitch.tv/t/update-improved-chat-server-restarts/1162/3

Twitch-IRC has the ability to reconnect gracefully upon server request.
Your current connection stay online for 30 seconds but there is another
connection under way, after 30 seconds, the old connection is dropped
and you continue with the new one.

We still need to wait for more informations from Twitch about this, but
we are ready to handle this kind of thing.